Improve the processing of Shield and Player modules
